Update pop_saveset.m
This line sometimes unexpectedly prevents a set from being saved, like whenever the EEG structure is modified via command line, for instance. While it does provide a warning, it is not easily visible. I think the risk of accidentally losing unsaved work is greater than the cost of saving again.